requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.120.4
favicon==0.7.0
favicorn==0.0.5
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
pyodbc==5.3.0
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
